Add language changer functionality
Issue: https://github.com/code4romania/seismic-risc/issues/97

This pr focuses on the language change functionality for our project.
I chose the Lingui Js library since it seems to offers the best features/ease of use.
One thing that I still have to do in order to complete this feature is to add documantation and a how to guide.
Right now the I pass the callbacks for changing the language using props. If we ever change to Redux or another state management system,
this can be improved.

Signed-off-by: Alexandru Buhai <alexandru.buhai@sectorlabs.ro>